requirements.txt
flask==3.1.0
flask-socketio==5.5.1
eventlet==0.38.2
python-dotenv==1.1.0
beautifulsoup4==4.13.4
pycryptodome==3.20.0
requests==2.32.4
tqdm==4.64.1
yt-dlp==2025.6.30
